Used short and unsigned in functions and variables

+++ Classes.h
@@ -3,35 +3,35 @@
 
 struct Ptrs {
     public:
-        int *ptrSpd;
-        int *ptrLuk;
-        int *ptrInt;
-        int *ptrWid;
-        int *ptrStr;
-        int *ptrChr;
+        unsigned short *ptrSpd;
+        unsigned short *ptrLuk;
+        unsigned short *ptrInt;
+        unsigned short *ptrWid;
+        unsigned short *ptrStr;
+        unsigned short *ptrChr;
 };
 
 class Class {
     public:
         Ptrs stats;
-        virtual int attack();
-        virtual int defense();
+        virtual unsigned int attack();
+        virtual unsigned int defense();
 };
 
 class Warrior: public Class {
     public:
-        int attack();
-        int defense();
+        unsigned int attack();
+        unsigned int defense();
 };
 
 class Hunter: public Class {
     public:
-        int attack();
-        int defense();
+        unsigned int attack();
+        unsigned int defense();
 };
 
 class Sorcerer: public Class {
     public:
-        int attack();
-        int defense();
+        unsigned int attack();
+        unsigned int defense();
 };
